+++ test/converter.ts
@@ -143,6 +143,15 @@ describe('Converter', () => {
       expect(disabled.html).toContain('&lt;i&gt;Hello!&lt;/i&gt;')
     })
 
+    it('correctly applies overridden global directives even if enabled HTML option', async () => {
+      const { rendered } = await instance({
+        html: true,
+        globalDirectives: { title: 'Hello' },
+      }).convert('<p>test</p>')
+
+      expect(rendered.title).toBe('Hello')
+    })
+
     it('strips UTF-8 BOM', async () => {
       const noBOM = await instance().convert('---\ntitle: test\n---')
       const BOM = await instance().convert('\ufeff---\ntitle: test\n---')